refactor: verify method moved to otp guard from otp service
