+++ lib/index.ts
@@ -1,4 +1,4 @@
 export * from './interfaces';
 export * from './otp.module';
-export * from './otp.service';
+export * from './services';
 export * from './guards';
